Update test and add fixture folder

+++ spec/services/weather_service_spec.rb
@@ -18,7 +18,7 @@ RSpec.describe WeatherService do
     expect(response[:current]).to have_key(:humidity)
     expect(response[:current][:humidity]).to be_a(Integer)
     expect(response[:current]).to have_key(:uvi)
-    expect(response[:current][:uvi]).to be_a(Integer)
+    expect(response[:current][:uvi]).to be_a(Float)
     expect(response[:current]).to have_key(:visibility)
     expect(response[:current][:visibility]).to be_a(Integer)
     expect(response[:current]).to have_key(:weather)
